remove unused methods
change structure to use packages
clean pom file
