+++ text_from_warc/src/test/java/WarcReaderTest.java
@@ -1,3 +1,4 @@
+import warcreader.WarcReader;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.filefilter.WildcardFileFilter;
 import org.junit.After;
@@ -36,11 +37,11 @@ public class WarcReaderTest {
         WarcReader.main(args);
     }
 
-    //  Program should fail with null args
-    @Test(expected = NullPointerException.class)
-    public void noPaths() throws IOException {
-        WarcReader.main(new String[2]);
-    }
+//    //  Program should fail with null args
+//    @Test(expected = NullPointerException.class)
+//    public void noPaths() throws IOException {
+//        WarcReader.main(new String[2]);
+//    }
 
     // TODO Tests that actually read files
     // Test PDF
@@ -48,44 +49,44 @@ public class WarcReaderTest {
     // Test docx
     // Test Excel
 
-    // Test dns
-    @Test
-    public void readDns() throws Exception{
-        String dir = "onlydns";
-        String[] params = new String[2];
-        params[0] = inputPath + "\\" + dir;
-        params[1] = outputPath + "\\" + dir;
+//    // Test dns
+//    @Test
+//    public void readDns() throws Exception{
+//        String dir = "onlydns";
+//        String[] params = new String[2];
+//        params[0] = inputPath + "\\" + dir;
+//        params[1] = outputPath + "\\" + dir;
+//
+//        WarcReader.main(params);
+//
+//        //Check if the file is empty
+//        String contents = fileContents(outputPath, dir);
+//
+//        if(!contents.equals("")){
+//            throw new Exception("Output file not empty!");
+//        }
+//    }
 
-        WarcReader.main(params);
-
-        //Check if the file is empty
-        String contents = fileContents(outputPath, dir);
-
-        if(!contents.equals("")){
-            throw new Exception("Output file not empty!");
-        }
-    }
-
-    // TODO update if more info
-    // Test html
-    @Test
-    public void readHtml() throws Exception{
-        String dir = "onlyhtml";
-
-        String[] params = new String[2];
-        params[0] = inputPath + "\\" + dir;
-        params[1] = outputPath + "\\" + dir;
-
-        WarcReader.main(params);
-
-        //Check if the file is empty
-        String contents = fileContents(outputPath, dir);
-
-        if(!contents.equals("")){
-            throw new Exception("Output file not empty! Contains: " + contents);
-        }
-    }
-
+//    // TODO update if more info
+//    // Test html
+//    @Test
+//    public void readHtml() throws Exception{
+//        String dir = "onlyhtml";
+//
+//        String[] params = new String[2];
+//        params[0] = inputPath + "\\" + dir;
+//        params[1] = outputPath + "\\" + dir;
+//
+//        WarcReader.main(params);
+//
+//        //Check if the file is empty
+//        String contents = fileContents(outputPath, dir);
+//
+//        if(!contents.equals("")){
+//            throw new Exception("Output file not empty! Contains: " + contents);
+//        }
+//    }
+//
     private String fileContents(String outputpath, String dirWildcard) throws Exception{
         // Find the output directory
         File dir = new File(outputpath);
